Reorganize DataAccess namespace
Mmoved MemoryDataAccessLayer to InMemory namespace, removed Csv namespace.
