+++ FireMothServices.Tests/DataAccess/Csv/MemoryDataAccessLayerTests.cs
@@ -14,9 +14,9 @@ using FluentAssertions;
 using Moq.AutoMock;
 using Xunit;
 using RiotClub.FireMoth.Services.DataAccess;
-using RiotClub.FireMoth.Services.DataAccess.Csv;
 using RiotClub.FireMoth.Services.Repository;
 using RiotClub.FireMoth.Services.Tests.Helpers;
+using RiotClub.FireMoth.Services.DataAccess.InMemory;
 
 public class MemoryDataAccessLayerTests
 {
